Remove the external call to check for paused state in FiatTokenUtil
The contract is almost never paused, so making an external call to
check at the beginning of the function is a waste of gas, since the
transferWithAuthorization call will end up checking whether the
contract is paused down the line anyway.

It is also recommended that users check for paused state with eth_call
before submitting any transaction.